+++ test/helper.rb
@@ -175,4 +175,16 @@ module ARHPTestSetup
   ensure
     mod.define_method(method_name, method_body)
   end
+
+  def simulate_rails_app_active_record_railties
+    if ActiveRecord.version < Gem::Version.new('6.1') || ENV['LEGACY_CONNECTION_HANDLING'] == 'true'
+      # Necessary for testing ActiveRecord 6.0 which uses the connection
+      # handlers when clearing query caches across all handlers when
+      # an operation that dirties the cache is involved (e.g. create/insert,
+      # update, delete/destroy, truncate, etc.)
+      ActiveRecord::Base.connection_handlers = {
+        ActiveRecord::Base.writing_role => ActiveRecord::Base.default_connection_handler
+      }
+    end
+  end
 end
