+++ test/test_arhp_with_new_connection_handling.rb
@@ -14,31 +14,6 @@ if ActiveRecord.version >= Gem::Version.new('6.1') && !ActiveRecord::Base.legacy
       Phenix.burn!
     end
 
-    def test_models_with_matching_hosts_and_non_matching_databases_issue_exists_without_arhp_patch
-      # Remove patch that fixes an issue in Rails 6+ to ensure it still
-      # exists. If this begins to fail then it may mean that Rails has fixed
-      # the issue so that it no longer occurs.
-      without_module_patch(ActiveRecordHostPool::ClearQueryCachePatch, :clear_query_caches_for_current_thread) do
-        without_module_patch(ActiveRecordHostPool::ClearQueryCachePatch, :clear_on_handler) do
-          exception = assert_raises(ActiveRecord::StatementInvalid) do
-            ActiveRecord::Base.establish_connection(:test_pool_1_db_a)
-            ActiveRecord::Base.cache { Pool1DbC.create! }
-          end
-
-          cached_db = ActiveRecord::Base.connection.unproxied.instance_variable_get(:@_cached_current_database)
-          assert_equal("Mysql2::Error: Table '#{cached_db}.pool1_db_cs' doesn't exist", exception.message)
-        end
-      end
-    end
-
-    def test_models_with_matching_hosts_and_non_matching_databases_do_not_mix_up_underlying_database
-      # ActiveRecord 6.0 introduced a change that surfaced a problematic code
-      # path in active_record_host_pool when clearing caches across connection
-      # handlers which can cause the database to change.
-      # See ActiveRecordHostPool::ClearQueryCachePatch
-      ActiveRecord::Base.cache { Pool1DbC.create! }
-    end
-
     def test_correctly_writes_to_sharded_databases
       AbstractShardedModel.connected_to(role: :writing, shard: :shard_b) do
         ShardedModel.create!
